Cached list of all available tables

Cached list of all available tables

Usage

tables

Format

tables

A data frame with 2339 rows and 8 columns:

id

Unique ID of the table

text

Description of the table

unit

The unit of the values represented in the table

updated

When this table was last updated

firstPeriod

The time of the first observation in the table

latestPeriod

The time of the latest observation in the table

active

Binary indicator of the table being active

variables

List of the variables in the table